Unit move queue.

+++ Assets/Scripts/Data/Stats/Stat.cs
@@ -8,28 +8,28 @@ using UnityEngine;
 public class Stat {
 
 	[SerializeField]
-	private int baseValue;	// Starting value
+	private float baseValue;	// Starting value
 
 	// List of modifiers that change the baseValue
-	private List<int> modifiers = new List<int>();
+	private List<float> modifiers = new List<float>();
 
 	// Get the final value after applying modifiers
-	public int GetValue ()
+	public float GetValue ()
 	{
-		int finalValue = baseValue;
+		float finalValue = baseValue;
 		modifiers.ForEach(x => finalValue += x);
 		return finalValue;
 	}
 
 	// Add new modifier
-	public void AddModifier (int modifier)
+	public void AddModifier (float modifier)
 	{
 		if (modifier != 0)
 			modifiers.Add(modifier);
 	}
 
 	// Remove a modifier
-	public void RemoveModifier (int modifier)
+	public void RemoveModifier (float modifier)
 	{
 		if (modifier != 0)
 			modifiers.Remove(modifier);
